Features include
- LBVH based tet level broad phase
- Code for timing and benchmarking to compare with CPU
- Plotting scripts for benchmarking results
- More performant Narrow Phase because of removal of complicated
  indexing

+++ geometry/proximity/sycl/utils/sycl_hydroelastic_surface_creator.h
@@ -17,7 +17,7 @@ inline SYCLHydroelasticSurface CreateHydroelasticSurface(
     const double* compacted_polygon_g_M, const double* compacted_polygon_g_N,
     const GeometryId* compacted_polygon_geom_index_A,
     const GeometryId* compacted_polygon_geom_index_B,
-    const size_t total_polygons) {
+    const uint32_t total_polygons) {
   // Transfer data from device to host
   std::vector<Vector3<double>> host_centroids(total_polygons);
   std::vector<double> host_areas(total_polygons);
